[#55] refactor : CouponUseHistoryCommand api 모듈 -> coupon 모듈로 이동
