+++ goodchoice-api/src/test/java/com/flab/goodchoiceapi/coupon/application/FakeCouponUseService.java
@@ -4,8 +4,10 @@ import com.flab.goodchoiceapi.coupon.infrastructure.*;
 import com.flab.goodchoiceapi.member.application.MemberQuery;
 import com.flab.goodchoicecoupon.application.CouponIssueCommand;
 import com.flab.goodchoicecoupon.application.CouponIssueQuery;
+import com.flab.goodchoicecoupon.application.CouponUseHistoryCommand;
 import com.flab.goodchoicecoupon.infrastructure.FakeCouponIssueCommand;
 import com.flab.goodchoicecoupon.infrastructure.FakeCouponIssueQuery;
+import com.flab.goodchoicecoupon.infrastructure.FakeCouponUseHistoryCommand;
 import com.flab.goodchoicecoupon.infrastructure.repositories.CouponIssueRepository;
 import com.flab.goodchoicecoupon.infrastructure.repositories.CouponUseHistoryRepository;
 import com.flab.goodchoicemember.infrastructure.repositories.MemberRepository;
