+++ goodchoice-coupon/src/main/java/com/flab/goodchoicecoupon/application/CouponUseHistoryCommand.java
@@ -1,4 +1,4 @@
-package com.flab.goodchoiceapi.coupon.application;
+package com.flab.goodchoicecoupon.application;
 
 import com.flab.goodchoicecoupon.domain.CouponUseHistory;
