+++ goodchoice-coupon/src/main/java/com/flab/goodchoicecoupon/infrastructure/CouponUseHistoryCommandImpl.java
@@ -1,16 +1,20 @@
-package com.flab.goodchoiceapi.coupon.infrastructure;
+package com.flab.goodchoicecoupon.infrastructure;
 
-import com.flab.goodchoiceapi.coupon.application.CouponUseHistoryCommand;
+import com.flab.goodchoicecoupon.application.CouponUseHistoryCommand;
 import com.flab.goodchoicecoupon.domain.CouponUseHistory;
 import com.flab.goodchoicecoupon.infrastructure.entity.CouponEntity;
 import com.flab.goodchoicecoupon.infrastructure.entity.CouponUseHistoryEntity;
 import com.flab.goodchoicecoupon.infrastructure.repositories.CouponUseHistoryRepository;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
-public class FakeCouponUseHistoryCommand implements CouponUseHistoryCommand {
+@Transactional
+@Component
+public class CouponUseHistoryCommandImpl implements CouponUseHistoryCommand {
 
-    private CouponUseHistoryRepository couponUseHistoryRepository;
+    private final CouponUseHistoryRepository couponUseHistoryRepository;
 
-    public FakeCouponUseHistoryCommand(CouponUseHistoryRepository couponUseHistoryRepository) {
+    public CouponUseHistoryCommandImpl(CouponUseHistoryRepository couponUseHistoryRepository) {
         this.couponUseHistoryRepository = couponUseHistoryRepository;
     }
